requirements.txt
click      7.1.2
pip        21.0.1
PyYAML     5.4.1
setuptools 54.2.0
wheel      0.36.2
